foundry_evm_coverage

Module analysis

Structs

ContractVisitor 🔒

A visitor that walks the AST of a single contract and finds coverage items.

SourceAnalysis

Coverage source analysis.

SourceFile

The source code and AST of a file.

SourceFiles

A list of versioned sources and their ASTs.

Functions

has_statements 🔒

Helper function to check if a given node is or contains any statement.